Realizo correcciones al codigo y se renombra la carpeta ServiceImpl

+++ src/main/java/com/nodo/retotecnico/service/BuysService.java
@@ -6,6 +6,6 @@ public interface IBuysService {
 
 public interface BuyService{
     List<Buy> getAllBuys();
-    Buy getBuyById(Long id);
+    Buy getBuyById(Integer id);
 
 }
